refactor: decompose agent_pane.rs (3,045 lines) into agent_pane/ module directory
Split the monolithic crates/phantom-app/src/agent_pane.rs into eight
focused sub-modules so parallel agents can work without merge conflicts:

  mod.rs          — AgentPane struct, type aliases, constants, constructors,
                    poll(), accessors, spawn_with_opts
  capture.rs      — flush_capture_record, push_snapshot, rollback_if_dirty
  dispatch_ctx.rs — build_dispatch_context, build_codebase_context,
                    class_label, now_unix_ms
  events.rs       — build_blocked/denied payloads, maybe_emit_*
  execute.rs      — dispatch, execute_pending_tools, format_tool_args
  journal.rs      — open_agent_journal
  spawn.rs        — impl App spawn methods + drain_blocked_events
  tests.rs        — all unit tests (358 pass)

Also fixes three pre-existing build errors in phantom-app:
  - Missing BrainConfig::catalog field in app.rs and boot_smoke.rs
  - Private field access (.id/.status/.output_log) using method calls
    in tests that broke when moved out of the inline mod scope

No behavior changes; all public API surfaces are preserved.
