+++ crates/phantom-app/tests/boot_smoke.rs
@@ -136,6 +136,7 @@ fn brain_spawns_and_channel_is_open() {
         enable_memory: false,
         quiet_threshold: 1.0, // suppress all actions so the thread stays quiet
         router: None,
+        catalog: None,
     });
 
     // The channel must be open immediately after spawn.
@@ -195,6 +196,7 @@ fn boot_smoke_full_invariants() {
         enable_memory: false,
         quiet_threshold: 1.0,
         router: None,
+        catalog: None,
     });
 
     assert!(
